auth/requirements.txt
fastapi==0.103.2
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.4.2
python-jose==3.4.0
python-multipart==0.0.20
requests==2.28.1
sqlalchemy==2.0.39
sqlalchemy-utils==0.41.2
uvicorn[standard]==0.23.2